pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic==0.40.0",
    "beautifulsoup4==4.12.3",
    "numpy==2.1.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "scikit-learn==1.5.2",
]
